Avoid using wildcards for checking filename extensions
Document what the blacklist is useful for, that it keeps editor's temp
files out of uploaded add-ons, and add Git's .orig files to the list.

According to Valgrind's Callgrind tool, around 2.7% of the time for
running unit tests is spent on checking for blacklisted filenames in
filesystem::is_legal_filename. This drops that to around 0.8%.