WIP: Add team_query wrapper for the teams array
This isn't really "in progress", it's just a rebase to remove the
already-merged changes to the add_source_file script.

WML numbers sides starting from 1, C++ likes indicies to start from zero, and so dealing
directly with a vector of sides requires handling the off-by-one numbering. This wrapper
encapsulates that logic.

This is a draft, looking at whether it would remove knowledge of the off-by-one
quirk from a sufficient number of places in the source tree to be useful. It's a
suggestion for reducing the number of files involved in PR 8271.

A wrapper for the sides or teams array, answering questions such as "are these units on
the same side".