Enable bugprone-move-forwarding reference
A forwarding reference is one of the form (note the template parameter):

    template<class X>
    void foo( X&& x );

The difference between this and a regular rvalue reference (where X is
not a template) is that foo() also accept lvalues. So it is possible to
call:

    std::string s = "Example";
    foo(s)

And if the body of foo(X&& x) std::moves x, then s will be invalidated.

This check warns on usage of std::move in this context and recommends
std::forward, which automatically determines whether to std::move
depending on whether foo() is called with an rvalue or an lvalue.