Create in [filter_ability] an attribute 'active' to replace _active suffix in [filter_ability_active]
Resolves https://github.com/wesnoth/wesnoth/issues/7926

Under pressure from @soliton-, I agree to merge [filter_ability] and [filter_ability_active] into a single filter equipped with the 'active' attribute which when the value is 'true' checks the units affected by the sought ability and when the value is 'false' will check the units carrying the ability even if they are not affected by it.

Because of the attribute strategy, I'm afraid that the developer will be mistaken about its function and that's why I'm not at all keen on it but I want to get out of this impasse due to the stubbornness of @soliton- (is it so bad to keep two tags like for ability_type and ability_type_active? there it would be clearer for those who manipulate these attributes because of the resemblance)