Lobby and game creation chatbox
Fixed no feedback when whispering to non existing user. The source of the error was trying to add message to the current whisper room with the add_chat_room_message_received function which is not meant to be used with whisper rooms. The error was solved by checking if the current room is whisper room and in this case selecting the lobby. There was an other error because during staging there is no lobby. This was solved by checking if the lobby room exists and if it doesnt then sending the message to the 'this game' room. Furthermore when receiving the cant find user message from the server the associated whiser room is closed.