wesnothd: handle server IP bans more generically
Instead of simply returning a string, `is_ip_banned` now returns a struct with an error code, ban reason, and ban time remaining. This avoids doing time duration formatting on the server and allows the error message to be localized on the client. It also makes the ban handling interface more generic in server_base, which should hopefully allow forum bans to be handled this way as well.

Drops the tor IP list stuff from wesnothd, since it doesn't fit well into this new paradigm.