Make variables in macros const
These macros cannot be autofixed due to the way clang-tidy functions.

Specifically, if the macro declares a variable that isn't changed, I
added a const declaration to the macro. If the macro is function-like
and has a variable that isn't changed in the macro, I also declared that
variable to be const.