Game Board: remove non-const overload of get_visible_unit
This was unique to this class, since the base class only implemented the const version. It was only needed for the whiteboard, and on closer inspection none of of the mutable references therein actually needed to be mutable. Marked them all const and removed the offending getter from game_board.

Note that the non-const overload had the same functionality as the const overload in the base class, only implemented via `find_visible_unit`.