Refactor unit bar drawing
This fixes an issue where the bar borders would always be one pixel wide, regardless of the zoom level or game render scaling. This made the bars look evidently procedurally drawn.

This is also an experiment with using normalized, fractional coordinates for rendering instead of absolute screen coordinates. I'd like to take this approach in more places, so this is a good proof of concept.