Add the PO files for wesnoth-tdg
Many of the CI builds expect the .po files to already exist, and
fail without them.

For SCons builds, the .pot file needs to exist, because po/SConscript
uses the existing files to find out which textdomains it needs to
update.

It was already added for CMake, but keeping the list in alphabetical
order makes maintenance easier.