Add macro for multiracial messages
It displays message depending on the race of the unit. Dwarves, ogres,
and undead use their own custom messages, the rest use the default one.

Oprional arguments allow changing the speaker and the image. By default,
narrator speaks for the undead, other units speak themselves.

Another optional argument is for the sound (same for all races).

This new macro is used for messages in scenarios 8 and 9, with
placeholder messages to be filled in later commit.