add a tag [affect_distant] to abilities
like [affect_adjacent] this tag affects other units than the owner of the ability but this time, it concerns units located in a radius defined in the tag or all units on the map.

I had already tried to make one a few years ago but the code was too heavy and I had given up.