Bump min and default game window size
Hopefully, we'll be able to implement better support for layouts that don't relay on pixel canvas sizes, but for now, this solves two important issues:

Using 1280x720 as the default window size means 3x auto pixel scaling for 4k fullscreen, which looks utterly horrendous and unusable. 2x is the practical limit of usability without going so far in the opposite direction as to actually reduce the available space.

Bumping the min window size also resolves #9898 and saves us from having to deal with layouts on unusably small windows.

As a side note, we should not be thinking in terms of pixel size anymore. I'm hoping SDL3 can give us better control over our HDPI support (which is much improved from before), but the fact remains that our current pixel scaling implementation is, essentially, a method to scale up the canvas if your resolution is grater than 1080p. It does nothing for size calculations, and it does nothing if you actually set a window size smaller than 1080p even on a 4k monitor.

So, for now, imperfect solution while we wait for SDL3, but solves the immediate issues with scaling and small window sizes.