TDG: remove silent raven speed reduction, and use TRAIT_SLOW instead
TDG gives Delfador access to a loyal raven/dire raven/war harbinger. Currently, the speed for this unit is capped at 6, to prevent it from being overly-powerful when compared with other options.

revolting_peasant pointed out that this speed cap makes the raven's stats inconsistent with the numbers in "view unit type". To resolve that, this PR replaces the speed cap with {TRAIT_SLOW}.