WFL/Formula: use a unique_ptr for the managed instance
Turns out we can use a unique_ptr after all, we just need to define the dtor out-of-line once function_symbol_table is a complete type.

Also adds a move ctor and assignment operator. This is necessary since the unique_ptr makes the class non-copyable, and also fixes a case in unit_filter_attribute_literal where a move was attempted but no move ctor provided.