Use three max_radius variables instead of one
With a single variable, if an ability has the tag [affect_distant]radius=5, then even abilities of a different type will search for units with affect_distant within a radius of 5 around the unit in action. To remedy this, I use a std::map variable for the radius of 5; this only applies to [leadership] abilities, for example, if one of them has [affect_distant]radius=5.

For filters and animations, the checked abilities can be of multiple types, so the variable is a single one.

Another separate variable is used for abilities with halo/overlay_image so that the search for units with [affect_distant] in this function only occurs if there is an ability with both halo/overlay_image and [affect_distant].