Generalise log domain matching rules
Previously, `--log-debug='*'` would not match all domains.
Similarly, `conf*` would not be a correct pattern, etc.

With this change, special handling of top-level domains is removed.
For example, the config domain will not be matched by `conf*`.

All domains can be matched by the magic string "all" as before,
or by using `*`.

Basic manual testing was performed for this PR:

* broken domain patterns that do not match anything
result in an error, as expected.

* patterns containing a comma (delimiter)
are handled "correctly", that is, individual components
are understood and processed separately.

* patterns "all" and "*" were tested.

* specifying "conf*" correctly enables the config domain.

* specifying "gui/*" does not enable the config domain.

Fixes #10052