TDG S11: fix last breath bug
In this scenario, the player wins during a "attack end" event. The player is also supposed to lose during a "last breath" event, if one of their heroes dies.

Because the engine triggers "attack end" before "last breath", it's possible to win at the same time one of your heroes dies, resulting in that hero being missing next scenario.

This fixes that bug.