help: Implement <img> scale= attribute and use it for unit portraits
Firstly, this patch adds an <img> scale= attribute that allows the user
to express rendering intent for images that have been transformed via
~SCALE_INTO() and similar IPFs. When this attribute is used, the UI
shrinks the image's draw size accordingly so that the final result does
not get upscaled according to the screen's pixel scale *again*, which
can result in excessively blurry images on high pixel density screens
(e.g. Apple Retina).

Secondly, this makes the unit type page generator convey the scale
information in <img>, since it's the generator and not the rendering
code that's currently in charge of picking the correct numbers for
scaling portraits up according to the display device's properties.

This all fixes a major issue on high pixel density screens where the
previous fixes for portrait upscaling quality and extents would cause
the portraits to be upscaled again on render, resulting in incredibly
blurry results (undermining the intent of the original fix) and a
massive lack of room for the unit description text due to half the
dialog being occupied by the portrait.