remove extra spaces in src/
I'm happy to remove the extra spaces, but in this case, it's best to do it in all the source files to encourage the developer engines to standardize their code writing.

The problem is that checking file by file is a real nightmare, and I doubt this PR can be merged as is, given the time it's going to take.