Optimize legacy tooltips (resolves #9327)
This removes the explicit texture setup in the tooltip ctor (now handled when a tooltip is actually ready to be drawn), as well as reduce redundant texture creation once it actually is initialized. We don't need to set the clip rect directly; setting a height or width is sufficient, since that's what actually controls the maximum size passed to the internal text renderer (if unset, it falls back to clip rect).