Separate support for special weapons and filtering into separate functions.
This will allow for clear code reading for each type of function instead of using multitasking functions, and also allow for simultaneous filtering of 'special_id' and 'special_type'. Previously, the functions had to be called once per attribute. We had [filter_weapon]special_id_active=a and special_type_active=damage. This required a full check for special_id_active, then a repeat check for special_type_active, and the detection of each of them wasn't always done at the same point. Now, we'll have a simultaneous check where the detected special will necessarily be a [damage]id=a, and not an [attacks]id=a accompanied by a [damage].