Unit Type: delete copy, default ctors (#10246)
#10053 inadvertently added a variable which copied a unit_type. This would have been caught had the copy ctor been deleted
(as copy assignment already is). Since these aren't objects we'll be passing around, better to delete both to catch such cases in future. Ironically, the only place that actually used the unit_type copy ctor was the unit_type copy ctor itself (and the code for copying the gender variations looks broken anyway - it was checking whether the wrong variable was non-null!)

Also added a default definition of the move assignment operator (we already had a move ctor), and deleted the default ctor (unimplemented).