Add [defense] ability to be able to modify base cth while be overwritable by [chance_to_hit].
In cases like 'dazed' (UTBs and TdG), you must specify the CTHs that should not be altered for the unit's precision part. For its defense, you must repeat the first operation, but also add a [chance_to_hit]add=10 apply_to=opponent active for marsksman when the base CTH is greater than 60. This is annoying in situations where eras or other resources are involved, hence the idea of a 'defense' ability.