Editor/Controller: keep better track of active menu
Builds on #9429. This adds a proper NONE value for active_menu_ and ensures that it is only set to an actual value for the duration during which a menu is displayed.

Also fixes an issue where get_action_state could return a bad value on first run since active_menu_ used to be initialized to MAP.