Move in_context_menu out of controller_base
This was defined as virtual in controller_base, but the only place it was called in a derived class was editor_controller, which used the default implementation. Meanwhile, play_controller::hotkey_handler, a command_executor, declared its own function of the same name. Moved the function to command_executor since it's a base class to both hotkey_handler and editor_controller and can provide a unified interface.