Revert nonsensical behavior change from 3cc3c75c69b
A trailing newline is counted as one more line by pango layout calculations
(pango_layout_get_extents). Have not found a configuration option for this or
an explanation on why that makes sense. To not show empty trailing lines for
tooltips keep removing a trailing newline.

This could perhaps be handled for all text rendering instead...