Remove FormulaAI (#10323)
This has been superseded by the Lua AI API. Its uses in mainline were removed in #5630.

This removes the entirety of the FAI engine and any related formula support in the AI code.
It also removes the unused and unimplemented ability to execute arbitrary commands using the
in-game console, since this relied on the FAI engine and was unimplemented for the Lua one.