Fix halo and hides display bug in abilities (#10363)
When a unit with the halo_image attribute [affect_adjacent] but with affect_self=no moves, it is likely to be illuminated by this halo, as it finds itself next to a temporary unit that is almost identical except for its location. This requires slightly complicating the verification with id if a simple comparison of unit is not sufficient.

same for [filter_adjacent] and [filter], if a unit with hides ability is invisible only when another unit of same race is adjacent, when unitmoves, it is 'hidden' because' adjacent to itself.