Menu Events: rework recruit validation (#10366)
This splits `unit_helper::recruit_message` into several smaller functions. Previously, both `recruit` and `do_recruit` ran all validation checks, which was unnecessary (we know we have a valid unit type in `do_recruit`, for instance, or that it's a valid recruit).

Error handling is now done in two stages:
- `recruit` and `repeat_recruit` handle the pre-action checks. Can we afford the unit, is it a valid type, etc.
- `do_recruit` assumes we have a valid type and now checks the requested recruit location.

The interface is much simpler now. We avoid calculating whiteboard gold for every unit checked in `recruit`, and get rid of the mutable references being passed around everywhere which necessitated unnecessary copies and made it very unclear what was modifying what (it also meant the error message function had side effects!)

We now have:
- `planned_gold_spent`: gives our "whiteboard gold". Splitting this off avoids recalculating it for every recruit checked
- `check_recruit_purse`: performs the actual "have enough gold" check. I'll add a version for recall separately.
- `validate_recruit_target`: This is now handled in `do_recruit`. We know we have a valid type that we can afford, now let's check locations. The mutable location references have been removed. This function returns a tuple with any error message regarding the target hex, the valid target, and the valid source.

Also moves the `command_disabler` to `do_recruit`. I *think* this is the right place for it, given the context of the recall function.