Consolidate the checks for abilities used as weapons into two separate functions.
Now that support for Boolean specials is separated from filtering, it would be better if the checks for self_ and other_ were handled by a single function instead of copying the code for each of them to simplify the writing. Hence, the use of two functions: one for handling Boolean specials, which will be called twice, and one for the filter functions, which will be called four times in total.