Remove unnecessary use of blank-hex.png for add-ons
Almost reverts d9ef464b1e9446b3a53e2d3bd86635529c897df4's changes to
this file, although not the other parts of that commit. The empty
string will be used instead of a filename that doesn't exist, because
non-existent filenames cause the canvas class to generate error logs
which lack the useful information identifying which add-on caused them.

Canvas' image_shape class will handle empty strings and simply not
draw an image. It will log a gui/draw message at log-level debug, but
there are many messages at that log level, and there's a separate
discussion about removing that log.

There's one edge case for locally authored add-ons without icons. This
commit means they will have a blank space for the icon, which is the
same as for any other add-on without an icon, although it will give an
error-level log message. Locally authored add-ons normally get an
overlay via an Image Path Function, which works with blank-hex.png but
not with an empty string. The UI probably shouldn't allow uploading
without an icon anyway.

This tidies up an if() ... if() that should be an if() ... else if().
Thanks to Nanamuru for pointing that out, it's one of the bugs found
by static analysis and reported in #10386, but it lead to this larger
change.