Log at most one "image won't be drawn" error per canvas
This was triggering for a broken Data URI in the list of add-ons,
which declares itself as image/png but the data is image/webp.
The whole Data URI was getting logged on every draw() call.

For formulas, this means that only the first failing filename or
data URI will be reported, even if the formula changes to a
different image. It seems unnecessary to add more logic; if the
person who sees the log is the person who can fix the problem,
then they're likely to either see a common pattern in a set of
broken filenames, or to keep testing until everything is fixed.