For locally authored add-ons without icons, don't add an IPF
If it's a local add-on then it gets an overlay via an Image Path
Function, which works with blank-hex.png but not with an empty string.
The UI should probably not allow uploading without an icon anyway...

This commit means these are shown with a blank space where the icon
would be, the same as for any other add-on that lacks an icon.