apply the 'neutral' alignment to the fire scepter and align the power of the attacks with those of the scepter when it is held by Li'sar
I know I've already tried to sell the concept in another PR but here's an argument that might be more relevant
Konrad is alignment lawful and Li'sar neutral, but the Scepter of Fire is so powerful that for them the difficulty comes more from keeping control in combat than trying to go to the maximum unlike the sword which comes from training, or from the magical mastery of mage which also comes with years of practice. For me, the fact that the Scepter attack is of reduced power at level 1 comes from the fact that the wearer lacks confidence and unconsciously restrains the power of the artifact. In the end, I consider that the alignment does not come into play because Konrad should not be more comfortable during the day than at night to handle what is a monster of power.