Simplify and consolidate save loading paths (#10453)
While working on #10451, I noticed that errors for saves weren't being displayed in-game. That was because b007b5a8d09980e10c5e7192f5620319e0d88edf made it so we checked the save index for save validity in order to avoid loading the actual save file twice - once to validate it before load_game_exception was thrown and once in the titlescreen loop to actually load the damn thing.

All in-game loading (autosaves and interactive) now reads the save file to check its validity and surface any errors *before* throwing load_game_exception. The load_game_metadata  payload is likewise now expected to contain the parsed save config. The autosave loader path has been adjusted to pass the loaded payload along with its exception. (Thankfully, gfgtdf's work already made this trivial.) Conversely, the interactive load path in game_launcher will assert if load data has already been provided.

The titlescreen loop has been simplified. Previously, we had two "load" paths: The first through the interactive Load Game menu on the titlescreen, which would load the save, reload the game config, and update the gamestate config before dismissing the titlescreen and entering a game. The second was through a load game exception, which would skip the titlescreen by updating the gamestate config and immediately launching the game. Both paths have now been unified. In the former case, the save is selected and its config loaded (to allow for any errors), before passing off to the game launcher to apply said config and launch the game. This means we aren't invoking the loading screen twice.

game_launcher::goto_editor was using the load_data payload when it had no reason to do so. The only piece of info actually needed was the filename, then passed on to the editor. Replaced the jump-to-editor flag with an optional string. The string may be empty, but that's a valid value for the editor.

The `savegame::loadgame` class has been removed and its constituent functionality made free-standing. There was really no reason to be a class, and its nature as such only made tracking the provenance of data and who was modifying what and when extremely difficulty. We now have `load_interactive`, `load_interactive_for_multiplayer`, and `load_interactive_by_exception`. (The last simply wraps the first). Each creates and return (or throws) its own load_game_metadata object.

The other two members of the loadgame class were similarly unnecessary. The game_config_view reference mostly existed to be passed to the Load Game dialog (it had one use in show_difficulty_dialog). The former has been removed; the game_config_manager is now queried when creating the dialog rather than passing the object in. The latter was turned into a local variable.

Finally, the gamestate (saved_game) reference. This is now handled by savegame::set_gamestate in order to make it extremely clear when the gamestate is being modified. Likewise, we no longer manually set the gamestate's game_classification object as part of the validation checks. We can just as easily query the version directly rather than relying on a temp object which will be later overwritten by set_state. `load_interactive_for_multiplayer` does still create its own game_classification object, but in that case, we need it for two checks. Would be good to dispense with that at some point...

Several early return checks has been removed, notable "is manager null" and "is filename empty". Neither of these should arise in the regular course of events. show_difficulty_dialog now writes the new difficulty value directly to the load config rather than leaving that to be handled separately.