Fix animation glitches when terrain scrolls onto screen
When an animated hex scrolled back on screen, instead of jumping to
the current frame of the animation, the engine played any intermediate
frames at a fast forwarded speed, because it incremented
current_frame_key_ at most once per redraw. The fix for that is the
second of the two loops changed in this commit.

When an animated hex first appears on screen, a while loop ran with
one iteration per cycle, as if the animation had been running since
the clock's epoch. Prior to b9bcdd3, the clock epoch seemed to be when
the engine loaded, after that commit it's whatever steady_clock uses,
which is implementation defined but made the lag more noticeable. The
fix is to just calculate the number of cycles to skip.

This code assumes that get_animation_duration returns a non-zero
value. That's already handled for units in frame_builder and for
terrain in terrain_builder::tile::rebuild_cache().

For a really short animation, floor(time_to_skip / acceleration_)
might be zero, but that implies that the full animation cycle is less
than 16ms, which sounds like a problem with the animation itself.
This code will cope with it, and for such a short animation it seems
reasonable to show any frame at random.

Thanks to Durzi (mentos987) for reporting the issue and working on a
fix.