Make the Load Game dialog see other-version save files from 1.19
If there are save directories for other versions of Wesnoth, there's a
drop-down list in the Load Game dialog to see those other directories.

This adapts 1.18 for commit 80fe9ba84ddb507f0212c445e4af9bb163e94bb3,
so that 1.18 will see the save directory of 1.19.2 onwards.

Previously, 1.18 would show the directory that 1.19.0 and 1.19.1 used.