Make the Load Game dialog for 1.19 and 1.20 see each other's save files
If there are save directories for other versions of Wesnoth, there's a
drop-down list in the Load Game dialog to see those other directories.

That directory moved in 80fe9ba84ddb507f0212c445e4af9bb163e94bb3, this
adapts so that the feature will continue working when 1.19 is an old
version. The code is the same as 1.18's implementation, I think the
technical debt of hardcoding the directory name "sync" in this place
is balanced by this being a place where some cross-version debt is
to be expected.