MP: Fix lag during AI turns on 5P The Wilderlands
The filter on the unit_placed event didn't check if the unit was
already loyal, and after 4 turns, the first unit belonging to side 5
had over 600 of these [object]s.

Thanks to gfgtdf for pointing out that the macro's upkeep=loyal will
only last until the unit is rebuilt, so even the units created at the
start should use an [object].

Any Walking Corpses created via plague won't get the object, but it
will be added if they advance to Soulless.